chore: remove evo from docker image (GPL license)
Evo is used for comparing portal accuracy during development, and is not needed when deploying. Moving it into a separate script, with instruction to pip install evo manually if needed.

The reason is that evo is GPL so we cannot redistribute it under MIT license (e.g. pre-built docker images)
